refactor : Car 클래스 불필요한 변수 제거 및 playGame()의 자동차  List 및 초기화 로직 Vehicles 클래스로 분리
